better reply/nick/host matching, code cleanup
Match reply messages with either a : or a , and ignore any non-nick
chars surrounding the replied to nick. E.g. "(somenick): text" will also
match.

There is still an issue with replies in the form of "SomeNick text", but
I'm not sure anything can be done about that.

Match nick and host masks case insensitively.

Remove the join/part event code and use the irc_nick infolist instead to
get the host, if available. We still need to set up a /whois query to
get the host infos at some point for real hostmask support.

Correct the code for distinguishing between nick and host masks.